MAILBOX-279: Split MessageMapperTest into MessageWithoutAttachment and MessageMapperTest
